feat: added method implementation to `Tree` trait and moved outside of utils
