+++ zk_prover/src/merkle_sum_tree/tree.rs
@@ -1,5 +1,6 @@
-use crate::merkle_sum_tree::utils::verify_proof;
-use crate::merkle_sum_tree::{MerkleProof, Node};
+use crate::merkle_sum_tree::{Entry, MerkleProof, Node};
+use halo2_proofs::halo2curves::bn256::Fr as Fp;
+use num_bigint::BigUint;
 
 /// A trait representing the basic operations for a Merkle-Sum-like Tree.
 pub trait Tree<const N_ASSETS: usize, const N_BYTES: usize> {
@@ -16,10 +17,46 @@ pub trait Tree<const N_ASSETS: usize, const N_BYTES: usize> {
     fn nodes(&self) -> &[Vec<Node<N_ASSETS>>];
 
     /// Generates a MerkleProof for the user with the given index.
-    fn generate_proof(
-        &self,
-        user_index: usize,
-    ) -> Result<MerkleProof<N_ASSETS, N_BYTES>, &'static str>;
+    fn generate_proof(&self, index: usize) -> Result<MerkleProof<N_ASSETS, N_BYTES>, &'static str> {
+        let nodes = self.nodes();
+        let depth = *self.depth();
+        let root = self.root();
+
+        if index >= nodes[0].len() {
+            return Err("The leaf does not exist in this tree");
+        }
+
+        let mut sibling_hashes = vec![Fp::from(0); depth];
+        let mut sibling_sums = vec![[Fp::from(0); N_ASSETS]; depth];
+        let mut path_indices = vec![Fp::from(0); depth];
+        let mut current_index = index;
+
+        let leaf = &nodes[0][index];
+
+        for level in 0..depth {
+            let position = current_index % 2;
+            let level_start_index = current_index - position;
+            let level_end_index = level_start_index + 2;
+
+            path_indices[level] = Fp::from(position as u64);
+
+            for i in level_start_index..level_end_index {
+                if i != current_index {
+                    sibling_hashes[level] = nodes[level][i].hash;
+                    sibling_sums[level] = nodes[level][i].balances;
+                }
+            }
+            current_index /= 2;
+        }
+
+        Ok(MerkleProof {
+            leaf: leaf.clone(),
+            root_hash: root.hash,
+            sibling_hashes,
+            sibling_sums,
+            path_indices,
+        })
+    }
 
     /// Verifies a MerkleProof.
     fn verify_proof(&self, proof: &MerkleProof<N_ASSETS, N_BYTES>) -> bool
@@ -27,6 +64,42 @@ pub trait Tree<const N_ASSETS: usize, const N_BYTES: usize> {
         [usize; N_ASSETS + 1]: Sized,
         [usize; 2 * (1 + N_ASSETS)]: Sized,
     {
-        verify_proof(proof)
+        let mut node = proof.leaf.clone();
+
+        let mut balances = proof.leaf.balances;
+
+        for i in 0..proof.sibling_hashes.len() {
+            let sibling_node = Node {
+                hash: proof.sibling_hashes[i],
+                balances: proof.sibling_sums[i],
+            };
+
+            if proof.path_indices[i] == 0.into() {
+                node = Node::middle(&node, &sibling_node);
+            } else {
+                node = Node::middle(&sibling_node, &node);
+            }
+
+            for (balance, sibling_balance) in balances.iter_mut().zip(sibling_node.balances.iter())
+            {
+                *balance += sibling_balance;
+            }
+        }
+
+        proof.root_hash == node.hash && balances == node.balances
+    }
+
+    /// Returns the index of the user with the given username and balances in the tree
+    fn index_of(&self, username: &str, balances: [BigUint; N_ASSETS]) -> Option<usize>
+    where
+        [usize; N_ASSETS + 1]: Sized,
+    {
+        let entry: Entry<N_ASSETS> = Entry::new(username.to_string(), balances).unwrap();
+        let leaf = entry.compute_leaf();
+        let leaf_hash = leaf.hash;
+
+        self.nodes()[0]
+            .iter()
+            .position(|node| node.hash == leaf_hash)
     }
 }
