+++ zk_prover/src/merkle_sum_tree/utils/mod.rs
@@ -1,17 +1,11 @@
 mod build_tree;
-mod create_proof;
 mod csv_parser;
 mod generate_leaf_hash;
 mod hash;
-mod index_of;
 mod operation_helpers;
-mod proof_verification;
 
 pub use build_tree::{build_merkle_tree_from_leaves, compute_leaves};
-pub use create_proof::create_proof;
 pub use csv_parser::parse_csv_to_entries;
 pub use generate_leaf_hash::generate_leaf_hash;
 pub use hash::{poseidon_entry, poseidon_node};
-pub use index_of::index_of;
 pub use operation_helpers::*;
-pub use proof_verification::verify_proof;
